+++ dhfs-parent/server/src/main/java/com/usatiuk/dhfs/files/objects/Directory.java
@@ -59,7 +59,7 @@ public class Directory extends FsNode {
     }
 
     @Override
-    public long estimateSize() {
-        return _children.size() * 192L;
+    public int estimateSize() {
+        return _children.size() * 192;
     }
 }
