+++ dhfs-parent/server/src/main/java/com/usatiuk/dhfs/files/objects/File.java
@@ -41,7 +41,7 @@ public class File extends FsNode {
     }
 
     @Override
-    public long estimateSize() {
-        return _chunks.size() * 192L;
+    public int estimateSize() {
+        return _chunks.size() * 192;
     }
 }
